+++ abstractions/python/kiota/abstractions/native_response_handler.py
@@ -1,7 +1,7 @@
 from typing import Any, Callable, Dict, Optional, TypeVar, cast
 
 from .response_handler import ResponseHandler
-from .serialization import Parsable
+from .serialization import Parsable, ParsableFactory
 
 NativeResponseType = TypeVar("NativeResponseType")
 ModelType = TypeVar("ModelType")
@@ -16,10 +16,10 @@ class NativeResponseHandler(ResponseHandler):
     # The error mappings for the response to use when deserializing failed responses bodies.
     # Where an error code like 401 applies specifically to that status code, a class code like
     # 4XX applies to all status codes within the range if a specific error code is not present.
-    error_map: Dict[str, Optional[Callable[[], Parsable]]]
+    error_map: Dict[str, Optional[ParsableFactory]]
 
     async def handle_response_async(
-        self, response: NativeResponseType, error_map: Dict[str, Optional[Callable[[], Parsable]]]
+        self, response: NativeResponseType, error_map: Dict[str, Optional[ParsableFactory]]
     ) -> ModelType:
         self.value = response
         self.error_map = error_map