+++ abstractions/python/kiota/abstractions/request_option.py
@@ -4,6 +4,7 @@ from abc import ABC, abstractmethod
 class RequestOption(ABC):
     """Represents a request option
     """
+
     @abstractmethod
     def get_key(self) -> str:
         """Gets the option key for when adding it to a request. Must be unique